feat: 컨트롤러 세션 검증 로직

+++ controllers/posts-controller.js
@@ -1,9 +1,9 @@
 import Post from '../models/Post.js'
 
-function getPosts(req, res){
+async function getPosts(req, res){
     try {    
-        const posts = Post.loadPosts();
-
+        const posts = await Post.loadPosts();
+        
         if(!posts){
             res.status(404).json({message: "존재하지 않는 리소스입니다"});
         }
@@ -17,18 +17,17 @@ function getPosts(req, res){
     }
 }
 
-function getPostDetail(req, res) {
+async function getPostDetail(req, res) {
     try {    
         const postId = parseInt(req.params.post_id);
         
         // 조회수 증가
         const incrementResult = Post.incrementView(postId);
         if(!incrementResult) {
-            return res.status(404).json({message: "게시글을 찾을 수 없습니다"});
+            return res.status(403).json({message: "잘못된 요청입니다"});
         }
 
-        const posts = Post.loadPosts();
-        const post = posts.find((p) => p.id == postId);
+        const post = await Post.findById(postId);
         if(!post) {
             return res.status(404).json({message: "게시글을 찾을 수 없습니다"});
         }
@@ -67,15 +66,22 @@ function updatePost(req, res) {
     try {
         const postId = parseInt(req.params.post_id);
         const { title, content, image } = req.body;
+        const userId = req.session.userId;
+        const post = Post.findById(postId);
+
+        if (!post) {
+            return res.status(404).json({ message: "게시글을 찾을 수 없습니다." });
+        }
+
+        if (post.author_id !== userId) {
+            return res.status(401).json({ message: "권한 없음" });
+        }
 
         if (!title && !content) {
             return res.status(400).json({ message: '제목과 내용을 입력해주세요.' });
         }
 
         const updatedPost = new Post(postId, null, title, content, image).updatePost();
-        if (!updatedPost) {
-            return res.status(404).json({ message: "게시글을 찾을 수 없습니다." });
-        }
 
         return res.status(200).json({
             message: "게시글 수정 성공",
@@ -90,13 +96,23 @@ function updatePost(req, res) {
 function deletePost(req, res) {
     try {
         const postId = parseInt(req.params.post_id);
+        const userId = req.session.userId;
+        const post = Post.findById(postId);
 
-        const success = Post.deletePost(postId);
-        if (!success) {
-            return res.status(404).json({ message: '게시글을 찾을 수 없습니다.' });
+        if (!post) {
+            return res.status(404).json({ message: "게시글을 찾을 수 없습니다." });
         }
 
-        res.status(200).json({ message: '게시글 삭제 성공' });
+        if (post.author_id !== userId) {
+            return res.status(401).json({ message: "권한 없음" });
+        }
+
+        const result = Post.deletePost(postId);
+        if (!result){
+            return res.status(500);
+        }
+
+        return res.status(200).json({ message: '게시글 삭제 성공' });
     } catch (error) {
         return res.status(500).json({ error: error.message });
     }
